Torroid: MainWindow: Xaml: Set NavigationView's PaneDisplayMode to "Auto" - Add CompactModeThresholdWidth to "0"
The NavigationView adapts between LeftCompact, and Left as the window gets wider.
LeftCompact:
The pane shows only icons until opened and is positioned to the left of the content. When opened, the pane overlays the content.
Left:
The pane is expanded and positioned to the left of the content.

- ref: https://learn.microsoft.com/en-us/windows/apps/design/controls/navigationview#display-modes
and
https://learn.microsoft.com/en-us/windows/apps/design/controls/navigationview#adaptive-behavior